fix: ProjetCRUD 버그 제거

+++ src/main/java/Project/Teaming/Project/Entity/Project.java
@@ -1,6 +1,7 @@
 package Project.Teaming.Project.Entity;
 
 import jakarta.persistence.*;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -13,7 +14,7 @@ public class Project {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(length = 50, nullable = false)
     private String title; //제목
